+++ src/Services/Dominion/Actions/EspionageActionService.php
@@ -803,8 +803,12 @@ class EspionageActionService
         if ($assassinsKilled > 0) {
             $unitsKilled['assassins'] = $assassinsKilled;
             $dominion->military_assassins -= $assassinsKilled;
-            if ($blackGuard && $assassinsKilled > 1) {
-                $this->queueService->queueResources('training', $dominion, ['military_assassins' => floor(0.75 * $assassinsKilled)]);
+            if ($assassinsKilled > 1) {
+                if ($blackGuard) {
+                    $this->queueService->queueResources('training', $dominion, ['military_assassins' => floor(0.75 * $assassinsKilled)]);
+                } elseif ($type == 'hostile') {
+                    $this->queueService->queueResources('training', $dominion, ['military_spies' => floor(0.25 * $assassinsKilled)]);
+                }
             }
         }